requirements.txt
aiohttp-socks
beautifulsoup4
Flask-Cors
Pillow
pynat
PyYAML
requests
selenium
webdriver-manager
